+++ src/middleware.ts
@@ -10,7 +10,6 @@ export async function middleware(request: NextRequest) {
     pathname.startsWith('/api/teams-bot') ||
     pathname.startsWith('/api/auth') ||
     pathname.startsWith('/api/sendgrid') ||
-    pathname.startsWith('/docs') ||
     pathname === '/api/set-locale' ||
     pathname.startsWith('/api/admin/') ||
     pathname.startsWith('/unsubscribe')